+++ packages/abc/simple-table/interface.ts
@@ -91,9 +91,11 @@ export interface SimpleTableColumn {
    */
   renderTitle?: string;
   /**
-   * 列宽，例如：`10%`、`100px`
+   * 列宽（数字型表示 `px` 值），例如：`100`、`10%`、`100px`
+   *
+   * **注意：** 固定列不支持百分比
    */
-  width?: string;
+  width?: string | number;
   /**
    * 排序的默认受控属性
    * - 只支持同时对一列进行排序，除非指定 `multiSort`，建议后端支持时使用