+++ packages/abc/simple-table/simple-table-row.directive.ts
@@ -1,28 +1,32 @@
 import { Directive, Input, TemplateRef, OnInit } from '@angular/core';
 import { SimpleTableComponent } from './simple-table.component';
 
-@Directive({
-  selector: '[st-row]',
-})
+@Directive({ selector: '[st-row]' })
 export class SimpleTableRowDirective implements OnInit {
-  @Input('st-row') id: string;
+  @Input('st-row')
+  id: string;
 
-  @Input() type: 'title';
+  @Input()
+  type: 'title';
 
   constructor(
-    public templateRef: TemplateRef<any>,
-    private table: SimpleTableComponent,
+    private ref: TemplateRef<any>,
+    private parent: SimpleTableComponent,
   ) {}
 
   ngOnInit(): void {
-    const col = this.table._columns.find(w => w.render === this.id);
+    const { _columns, _customTitles, _customRows } = this.parent;
+    const col = _columns.find(w => w.render === this.id);
     if (!col) return;
+
     switch (this.type) {
       case 'title':
-        col.__renderTitle = this.templateRef;
+        col.__renderTitle = this.ref;
+        _customTitles[this.id] = this.ref;
         break;
       default:
-        col.__render = this.templateRef;
+        col.__render = this.ref;
+        _customRows[this.id] = this.ref;
         break;
     }
   }