+++ packages/abc/simple-table/simple-table.component.ts
@@ -68,22 +68,29 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
   _allChecked = false;
   _indeterminate = false;
   _columns: SimpleTableColumn[] = [];
+  _customTitles: { [key: string]: TemplateRef<any> } = {};
+  _customRows: { [key: string]: TemplateRef<any> } = {};
 
   //#region fields
 
   /** 数据源 */
-  @Input() data: string | any[] | Observable<any[]>;
+  @Input()
+  data: string | any[] | Observable<any[]>;
   /**
    * 额外请求参数，默认自动附加 `pi`、`ps` 至URL
    * - `{ status: 'new' }` => `url?pi=1&ps=10&status=new`
    */
-  @Input() extraParams: any;
+  @Input()
+  extraParams: any;
   /** 请求方法 */
-  @Input() reqMethod: string = 'GET';
+  @Input()
+  reqMethod: string = 'GET';
   /** 请求体 `body` */
-  @Input() reqBody: any;
+  @Input()
+  reqBody: any;
   /** 请求体 `Header` */
-  @Input() reqHeaders: any;
+  @Input()
+  reqHeaders: any;
   /**
    * 重命名请求参数 `pi`、`ps`
    * - `{ pi: 'Page' }` => `pi` 会被替换成 Page
@@ -97,7 +104,8 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
   }
   private _reqReName: ReqReNameType = { pi: 'pi', ps: 'ps' };
   /** 请求异常时回调 */
-  @Output() readonly reqError: EventEmitter<any> = new EventEmitter<any>();
+  @Output()
+  readonly reqError: EventEmitter<any> = new EventEmitter<any>();
   /**
    * 重命名返回参数 `total`、`list`
    * - `{ total: 'Total' }` => Total 会被当作 `total`
@@ -123,7 +131,8 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
   }
   private _resReName: ResReNameType = { total: ['total'], list: ['list'] };
   /** 列描述  */
-  @Input() columns: SimpleTableColumn[] = [];
+  @Input()
+  columns: SimpleTableColumn[] = [];
   /** 每页数量，当设置为 `0` 表示不分页，默认：`10` */
   @Input()
   get ps() {
@@ -188,9 +197,11 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
   }
   private _bordered = false;
   /** table大小 */
-  @Input() size: 'small' | 'middle' | 'default' = 'default';
+  @Input()
+  size: 'small' | 'middle' | 'default' = 'default';
   /** 纵向支持滚动，也可用于指定滚动区域的高度：`{ y: '300px', x: '300px' }` */
-  @Input() scroll: { y?: string; x?: string };
+  @Input()
+  scroll: { y?: string; x?: string };
   /** 是否显示pagination中改变页数 */
   @Input()
   get showSizeChanger() {
@@ -201,7 +212,8 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
   }
   private _showSizeChanger = false;
   /** pagination中每页显示条目数下拉框值 */
-  @Input() pageSizeOptions: number[];
+  @Input()
+  pageSizeOptions: number[];
   /** 是否显示pagination中快速跳转 */
   @Input()
   get showQuickJumper() {
@@ -254,7 +266,8 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
   }
   private _isPageIndexReset = true;
   /** 分页方向 */
-  @Input() pagePlacement?: 'left' | 'center' | 'right' = 'right';
+  @Input()
+  pagePlacement?: 'left' | 'center' | 'right' = 'right';
   /** 切换分页时返回顶部 */
   @Input()
   get toTopInChange() {
@@ -274,7 +287,8 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
   }
   private _toTopOffset = 100;
   /** 重命名排序值，`columns` 的重命名高于属性 */
-  @Input() sortReName: { ascend?: string; descend?: string };
+  @Input()
+  sortReName: { ascend?: string; descend?: string };
   /** 是否多排序，当 `sort` 多个相同值时自动合并，建议后端支持时使用 */
   @Input()
   get multiSort() {
@@ -296,18 +310,24 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
   }
   private _multiSort: boolean | SimpleTableMultiSort;
   /** 数据处理前回调 */
-  @Input() preDataChange: (data: SimpleTableData[]) => SimpleTableData[];
+  @Input()
+  preDataChange: (data: SimpleTableData[]) => SimpleTableData[];
   /** `header` 标题 */
-  @ContentChild('header') header: TemplateRef<void>;
+  @ContentChild('header')
+  header: TemplateRef<void>;
   /** 额外 `body` 内容 */
-  @ContentChild('body') body: TemplateRef<void>;
+  @ContentChild('body')
+  body: TemplateRef<void>;
   /** `footer` 底部 */
-  @ContentChild('footer') footer: TemplateRef<void>;
+  @ContentChild('footer')
+  footer: TemplateRef<void>;
   /** `expand` 可展开，当数据源中包括 `expand` 表示展开状态 */
   @ContentChild('expand')
   expand: TemplateRef<{ $implicit: any; column: SimpleTableColumn }>;
-  @Input() noResult: string | TemplateRef<void>;
-  @Input() widthConfig: string[];
+  @Input()
+  noResult: string | TemplateRef<void>;
+  @Input()
+  widthConfig: string[];
   /** 页码、每页数量变化时回调 */
   @Output()
   readonly change: EventEmitter<SimpleTableChange> = new EventEmitter<
@@ -324,7 +344,8 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
     SimpleTableData
   >();
   /** 排序回调 */
-  @Output() readonly sortChange: EventEmitter<any> = new EventEmitter<any>();
+  @Output()
+  readonly sortChange: EventEmitter<any> = new EventEmitter<any>();
   /** Filter回调 */
   @Output()
   readonly filterChange: EventEmitter<SimpleTableColumn> = new EventEmitter<
@@ -341,7 +362,8 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
     SimpleTableRowClick
   >();
   /** 行单击多少时长之类为双击（单位：毫秒），默认：`200` */
-  @Input() rowClickTime = 200;
+  @Input()
+  rowClickTime = 200;
   /** 后端分页是否采用`1`基索引，只在`data`类型为`string`时有效 */
   @Input()
   get zeroIndexedOnPage() {
@@ -751,12 +773,10 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
   }
 
   clearFilter() {
-    this._columns
-      .filter(w => w.filtered === true)
-      .forEach(col => {
-        col.filtered = false;
-        col.filters.forEach(f => f.checked = false);
-      });
+    this._columns.filter(w => w.filtered === true).forEach(col => {
+      col.filtered = false;
+      col.filters.forEach(f => (f.checked = false));
+    });
   }
 
   //#endregion
@@ -865,7 +885,7 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
   }
 
   _btnClick(e: Event, record: any, btn: SimpleTableButton) {
-    e.stopPropagation();
+    if (e) e.stopPropagation();
     if (btn.type === 'modal' || btn.type === 'static') {
       const obj = {};
       obj[btn.paramName || this.defConfig.modalParamsName || 'record'] = record;
@@ -918,15 +938,24 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
   //#region fixed
 
   private fixedCoerce(list: SimpleTableColumn[]) {
-    list.forEach((item, idx) => {
-      if (item.fixed && item.width) {
-        if (item.fixed === 'left') {
-          item._left = idx === 0 ? '0px' : list[idx - 1].width;
-        } else {
-          item._right = idx === list.length - 1 ? '0px' : list[idx + 1].width;
-        }
-      }
-    });
+    const countReduce = (a: number, b: SimpleTableColumn) =>
+      a + +b.width.toString().replace('px', '');
+    // left width
+    list
+      .filter(w => w.fixed && w.fixed === 'left' && w.width)
+      .forEach(
+        (item, idx) =>
+          (item._left = list.slice(0, idx).reduce(countReduce, 0) + 'px'),
+      );
+    // right width
+    list
+      .filter(w => w.fixed && w.fixed === 'right' && w.width)
+      .reverse()
+      .forEach(
+        (item, idx) =>
+          (item._right =
+            (idx > 0 ? list.slice(-idx).reduce(countReduce, 0) : 0) + 'px'),
+      );
   }
 
   //#endregion
@@ -1064,6 +1093,12 @@ export class SimpleTableComponent implements OnInit, OnChanges, OnDestroy {
       item.buttons = this.btnCoerce(item.buttons);
       // i18n
       if (item.i18n && this.i18nSrv) item.title = this.i18nSrv.fanyi(item.i18n);
+      // custom row
+      if (item.render) {
+        item.__renderTitle = this._customTitles[item.render];
+        item.__render = this._customRows[item.render];
+      }
+
       ++idx;
       newColumns.push(item);
     }